requirements.txt
streamlit==1.40.0
feedparser==6.0.10
PyGithub==2.1.1
google-generativeai==0.8.3
python-dotenv==1.0.0
